backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
httpx>=0.26.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
apscheduler>=3.10.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
